light-dark() does not work in SVGImages https://bugs.webkit.org/show_bug.cgi?id=283489 rdar://140661763
Reviewed by NOBODY (OOPS!).

Page::useColorAppearance() is not propagated to the Page of the SVGImage when it
is first created. To fix this issue, SVGImage will query the container Page
useSystemDarkAppearance() through the ImageObserver and set it in its Page.

Also if the dark mode changes after the SVGImage is loaded, the SVGImage does not
recalculate its style colors. To sync the setUseColorAppearance() of all the
SVGImages when the system appearance changes, Page::setUseColorAppearance() needs
to scan all the SVGImages in all Documents and call setUseColorAppearance() of
the Pages of these Documents.

* LayoutTests/platform/mac-wk1/TestExpectations:
* LayoutTests/svg/custom/resources/light-dark-red-green.svg: Added.
* LayoutTests/svg/custom/svg-image-dark-mode-dynamic-expected.html: Added.
* LayoutTests/svg/custom/svg-image-dark-mode-dynamic.html: Added.
* LayoutTests/svg/custom/svg-image-dark-mode-initial-expected.html: Added.
* LayoutTests/svg/custom/svg-image-dark-mode-initial.html: Added.
* Source/WebCore/loader/cache/CachedImage.cpp:
(WebCore::CachedImage::useSystemDarkAppearance const):
* Source/WebCore/loader/cache/CachedImage.h:
* Source/WebCore/loader/cache/CachedImageClient.h:
(WebCore::CachedImageClient::useSystemDarkAppearance const):
* Source/WebCore/page/Page.cpp:
(WebCore::Page::setUseColorAppearance):
* Source/WebCore/platform/graphics/ImageObserver.h:
(WebCore::ImageObserver::useSystemDarkAppearance const):
* Source/WebCore/rendering/RenderObject.cpp:
(WebCore::RenderObject::useSystemDarkAppearance const):
* Source/WebCore/rendering/RenderObject.h:
* Source/WebCore/svg/graphics/SVGImage.cpp:
(WebCore::SVGImage::dataChanged):