[Navigation API] navigate.back doesn't work in main frame when preceded by navigate.back in child frame (part 2) https://bugs.webkit.org/show_bug.cgi?id=297365 rdar://158259024
Reviewed by NOBODY (OOPS!).

(I'm putting this on EWS to see if any other tests break).

The test disambigaute-forward.html fails on this line:
"assert_equals(navigation.currentEntry.index, start_index);"
because the mainframe's back navigation fails.

This is what the test is doing:
1. Load a main frame and an iframe
2. Fragment navigate the main frame
3. Fragment navigate the iframe
4. Go back in the iframe (this succeeds)
5. Go back in the main frame (this fails to go back)

The main frame's m_entries correctly looks like:

MainFrame (URL: disambigaute-forward.html, itemID: A)
MainFrame (URL: disambigaute-forward.html#top, itemID: B)

But on the navigate.back, the item we're looking for is wrong.
We're looking for item B when we should have been looking for item A.

Why? On navigate.back, for most of the code path, we are correctly
using item A. But then:

FrameLoader::loadInSameDocument() ends up calling
HistoryController::recursiveUpdateForSameDocumentNavigation()
which sets the currentItem to the provisionalItem. The currentItem
is the item we're looking for. Up to this point, it has correctly
been item A. But here, we set it to item B.

Turns out, the provisionalItem got set to item B by the iframe's
back navigation in HistoryController::recursiveSetProvisionalItem().
In the iframe's back, we actually call recursiveSetProvisionalItem
from the main frame's history controller. This means that we set the
provisional item for the main frame--even though the main frame is not
actually navigating. Since the main frame isn't navigating, this
provisional item does not get set to nullptr when the iframe's load
finishes. So later on, this causes the main frame's back to
look for the wrong item and for the back to fail.

On the iframe's back, this provisional item should have been cleared
by recursiveUpdateForSameDocumentNavigation. But it doesn't get cleared
because shouldDoSameDocumentNavigationTo is false. The current item
and provisional item are both item B. This is because the main frame's
history controller's current item does not change when the iframe navigates.

So this check is preventing the main frame from navigating correctly and
cutting off the recursive update of the rest of the rest of the frames
as well.

I'm not sure if the issue is something else, but I'm trying out removing
this check to see if it breaks anything else.

* LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/disambigaute-forward-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/disambigaute-traverseTo-forward-multiple-expected.txt:
* Source/WebCore/loader/HistoryController.cpp:
(WebCore::HistoryController::recursiveUpdateForSameDocumentNavigation):