[build.webkit.org][ews-build] The return code from the command executed is not forwarded properly on step run-api-tests https://bugs.webkit.org/show_bug.cgi?id=297755
Reviewed by NOBODY (OOPS!).

The shell command added in 279232@main and 292611@main to save the logs on a file and then grep
for the final result is masking the return code of the command executed to run the tests.

The return code is always zero (assuming that grep can find the "Ran " string)

Fix that by saving the return code of the command to run the tests on a env variable and
returning that instead.

* Tools/CISupport/build-webkit-org/steps.py:
(RunAPITests.run):
* Tools/CISupport/build-webkit-org/steps_unittest.py:
* Tools/CISupport/ews-build/steps.py:
(RunAPITests.run):
* Tools/CISupport/ews-build/steps_unittest.py: