[Site Isolation] change downcast to dynamicDowncast in JSDOMBindingSecurity::remoteFrameAccessError https://bugs.webkit.org/show_bug.cgi?id=297874 rdar://159134034
Reviewed by NOBODY (OOPS!).

Change a downcast to a dynamicDowncast to handle cases where an invalid downcast between a RemoteDOMWindow to a LocalDOMWindow was attempted. Previously, if the cast failed, an assert was tripped and caused a crash.

See Bugzilla for steps to reproduce. The crash only occurred with Site Isolation enabled.

* LayoutTests/http/tests/site-isolation/window-properties-expected.txt:
* Source/WebCore/bindings/js/JSDOMBindingSecurity.cpp:
(WebCore::remoteFrameAccessError):
(WebCore::reportErrorAccessingRemoteFrame):
(WebCore::BindingSecurity::shouldAllowAccessToDOMWindow):