Convert PlatformEvent timestamps to MonotonicTime rdar://158717958
PlatformEvent objects represent user input such as taps, mouse clicks
or key presses, and are used to create used to create DOM UIEvent
objects. The DOM UIEvent class uses MonotonicTime as its time base, so
it is natural for PlatformEvents to do so as well.

This patch avoids an unnecessary round trip: timestamps were being
converted from native timestamps (monotonic) to WallTime, and later to
MonotonicTime. The conversion to WallTime introduces small errors
which cause inconsistencies when comparing performance.now() to event
timestamps.

This patch is not supposed to cause any change in behavior. However,
these two changes are not strictly refactors:

* Tools/WebKitTestRunner/mac/EventSenderProxy.mm : issued events now
  contain a timestamp with a small fixed value instead of zero. This is
  required to make sure the timestamp is not interpreted as a falsy
  value in some scrolling logic downstream - an event with a
  zeroed timestamp will produce a MonotonicTime with the falsy, default
  initialized value; this doesn't happen with WallTime.
* Source/WebCore/platform/ios/PlatformEventFactoryIOS.mm : the
  PlatformTouchEventBuilder(WebEvent *event) constructor used to create
  a WallTime object directly from an event timestamp, with
  WallTime::fromRawSeconds(event.timestamp); the change to
  MonotonicTime::fromRawSeconds(event.timestamp) changes how
  event.timestamp is interpreted (from seconds since 1970 to seconds
  since system boot). The previous behavior was possibly an error, and
  the only code path that uses the affected constructor is the touch()
  method in Source/WebKitLegacy/mac/WebView/WebHTMLView.mm.