proxmox.com: Large amounts of extra vertical spacing between content. https://bugs.webkit.org/show_bug.cgi?id=296988 rdar://157638931
Reviewed by NOBODY (OOPS!).

During track sizing we are supposed to treat any auto margins on grid
items as 0. These would then be resolved after we have computed the grid
area for the item. However, the attached testcase demonstrates that we
are not doing this correctly in all scenarios for at least the inline
margins.

The content is along the lines of
<grid>
  <subgrid columns style="margin-left: auto">
    <div> X X X X X X </div>
  </subrid columns>
</grid>

When we resolve the row sizes for the grid we end up calling layout() on
the subgrid via GridTrackSizingAlgorithmStrategy::logicalHeightForGridItem.
During the subgrid's layout while we are in updateLogicalWidth(), we end up
computing a logical width of 0 and a resulting margin-left value of
whatever the available space is. Then, we end up setting the subgrid's
columns to the column sizes of the parent grid since it is subgridded in
that dimension, but we also factor out the subgrid's incorrectly
computed margins which are the same size as the columns from the parent.
This ends up with column sizes of 0 for the content of the subgrid.
Later on when we compute the rows of the subgrid, we end up setting an
available space of 0 for its content which results in really tall
content and really tall rows in the parent as a result.

This patch corrects this behavior for the inline margins of grid item's
by keeping track of a new member on RenderGrid which keeps track of
whether or not we are in the middle of track sizing via SetForScope.
When we are resolving inline margins for grid items we set any auto
margins to 0 if we are in the middle of track sizing by checking for
this state.