Undo zoom application to rem and em units in SVG length resolution https://bugs.webkit.org/show_bug.cgi?id=298102 rdar://159448970
Reviewed by NOBODY (OOPS!).

Currently, when resolving SVGLengthValue, somewhere in the callstack
we call computeUnzoomedNonCalcLengthDouble to get user units as the
final value. This value can sometimes be resolved against the computed
font size, which may or may not include zoom as when units are rem/em.

In SVG contexts when using rem/em, we expect these values to be unzoomed
because SVG applies zoom during layout-time at viewport level.
(see RenderSVGViewportContainer)

Ideally, a routine named "computeUnzoomed*" should guarantee
non-zoomed values, but this isn't the case. We can either patch the
other side (computeUnzoomedNonCalcLengthDouble) to always resolve
against the specificed size, or fix this side (SVGLengthContext).
I chose this side.

This PR undoes the applied zoom for rem and em units in SVG length context.

* LayoutTests/TestExpectations:
* Source/WebCore/svg/SVGLengthContext.cpp:
(WebCore::SVGLengthContext::computeNonCalcLength const):