Click and auxclick event targeting does not follow pointer capture target override https://bugs.webkit.org/show_bug.cgi?id=298130 rdar://159477637
Reviewed by NOBODY (OOPS!).

When an element has pointer capture, [aux]click events should (also) be
dispatched to the pointer capture target override rather than the hit
tested node under the pointer, refer to
https://www.w3.org/TR/pointerevents/#:~:text=even%20though%20the%20lostpointercapture%20event%20with%20the%20same%20pointerId%20has%20been%20dispatched%20already

In this patch, we add a new m_clickCaptureElement member to
EventHandler. This stores the element when pointer capture is
established during a pointer event sequence. Note that we choose to not
override m_clickNode because it is used for general event dispatch
logic, too.

The swallowAnyClickEvent() method now checks for this captured element
first when determining the target for click events.

The patch fixes (and updates expectations for) the failing tests:
- pointerevent_click_during_capture_mouse-click.html
- pointerevent_click_during_capture_mouse-auxclick.html

* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_click_during_capture_mouse-auxclick-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_click_during_capture_mouse-click-expected.txt:
* Source/WebCore/page/EventHandler.cpp:
(WebCore::EventHandler::clear):
(WebCore::EventHandler::nodeWillBeRemoved):
(WebCore::EventHandler::invalidateClick):
(WebCore::EventHandler::swallowAnyClickEvent):
(WebCore::EventHandler::pointerCaptureElementDidChange):
* Source/WebCore/page/EventHandler.h: