[Swift in WebKit] Add a modulemap for the private interface of JSC https://bugs.webkit.org/show_bug.cgi?id=298124 rdar://159472949
Reviewed by NOBODY (OOPS!).

Add a private module map to JSC, a prerequisite for modularizing WebCore.

* Source/JavaScriptCore/API/JSContextInternal.h:
* Source/JavaScriptCore/API/JSContextPrivate.h:
* Source/JavaScriptCore/API/JSScript.h:
* Source/JavaScriptCore/API/JSValueInternal.h:
* Source/JavaScriptCore/API/JSValuePrivate.h:
* Source/JavaScriptCore/API/JSVirtualMachinePrivate.h:

Add some missing OBJC guards, so that JSC can be compiled in C++ only mode.

* Source/JavaScriptCore/Configurations/JavaScriptCore.xcconfig:

- Ensure the module verifier runs on the public SDK to make it easier to catch issues
- Adjust the module verification flags to use the built WTF/bmalloc modules instead of the system ones
- Disable verification for Obj-C and C. This is done because the JSC private interface would require
the addition of __cplusplus guards to every private header file. This itself wouldn't be needed usually,
however the module verifier does not support module require clauses; if this functionality becomes supported,
or if a custom verifier is written, these languages can be re-added. It is reasonably safe to omit this
subset of languages since ObjC++ and C++ are still covered, and it is unlikely for C++ API to accidentally
be added to the public JSC API.

* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/JavaScriptCore_Private.modulemap: Added.