[MSE][GStreamer] Initial support for multiple video tracks on MSE https://bugs.webkit.org/show_bug.cgi?id=298245
Reviewed by NOBODY (OOPS!).

Multiple video tracks in MSE are (optionally) supported by the spec[1],
but have never been used in practice in WebKitGTK/WPE. Creating more
than one video SourceBuffer causes a not-linked pipeline error in
WebKitMediaSrc. It would be nice to have a basic support of that use
case, since it's apparently required for the YouTube EME tests.

See: https://github.com/WebPlatformForEmbedded/WPEWebKit/issues/1549

This patch sanitizes the video track selection state on creation (only
one video track can be active at once, all of them appeared as active on
creation before), mirrors the selection state from the video tracks
known to the HTMLMediaElement to the ones knows by the player private
(they're a different set in the case of MSE), and adds support for
unregistered streams in WebKitMediaSrc (the ones causing the not-linked
error in the first place), just avoiding pushing buffers to pads with
unlinked steams.

This patch only provides an initial support. It doesn't address complex
scenarios such as the init segment and data reenqueues that should be
needed if the switch happens after the init segment and data have been
appended (and enqueued) and the track selection change happens after
that. In such cases, the pipeline is refactored but, since the needed
data hasn't reached the video decoder on time (because the decoder just
didn't exist at enqueue time), only future samples will be enqueued from
that time on and decoding artifacts will happen.

[1] https://w3c.github.io/media-source/#dfn-sourcebuffer-configuration

* Source/WebCore/Modules/mediasource/MediaSource.cpp:
(WebCore::MediaSource::addVideoTrackToElement): Select the first video track added and unselect the next ones to ensure that only one video track is active at a given time after adding several ones.
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::mirrorEnabledVideoTrackIfNeeded): Propagate the "selected" state from a given VideoTrackPrivateGStreamer (typically one know by HTMLMediaElement) to the internal VideoTrackPrivateGStreamer managed by the player private (if different from the supplied one) having the same track ID.
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.h: Added mirrorEnabledVideoTrackIfNeeded().
* Source/WebCore/platform/graphics/gstreamer/VideoTrackPrivateGStreamer.cpp:
(WebCore::VideoTrackPrivateGStreamer::setSelected): Propagate the "selected" state to the mirror track that the player private may have.
* Source/WebCore/platform/graphics/gstreamer/mse/WebKitMediaSourceGStreamer.cpp:
(WebKitMediaSrcPrivate::streamById): Support asking for streams not registered in this element. Log an info instead of asserting when not present.
(webKitMediaSrcTearDownStream): Gracefully bail out if the Stream is not registered in this element.
(webKitMediaSrcLoop): Ditto. Just skip the buffer push without triggering an error.
(webKitMediaSrcFlush): Ditto.